requirements.txt
arxiv==2.1.0
crossref-commons==0.0.7
trafilatura==1.9.0
beautifulsoup4==4.12.3
requests
pydantic
python-dotenv
chromadb
typer
rich
PyYAML
fastapi
uvicorn
pypdf
lxml>=4.9.3
httpx>=0.27.0